[test] merge testtheme.go into theme.go
